Add header component (navbar)
This header component will either contain a link to the login page, or a
logout button. The link/button will be dynamically displayed based on if
a non-expired JWT is present in local storage or not.

I haven't added this new logic yet, this commit only holds the initial
Login component.