Implement page size settings
I've also changed up the pagination to pre-fetch all records. This will
make it easier for me to do sorting and searching when I implement that
-- I don't want to sort or search on only a page of data, and updating
the endpoint to allow for more sophisticated filtering seems harder at
the moment than implementing all that functionality on the front-end.

If the page becomes slow enough that it's not worth the trade-off, I'll
revisit handling that functionality on the server-side and fetching data
one page at a time.