Implement "admin" mode on history page
When in admin mode, a user will be able to delete transactions in the
transaction history table. This isn't a feature I'd normally allow a
calculator user to make use of, but I think it makes sense for an
administrator role.

Since this project isn't going to get any serious use, I will allow any
user to toggle admin mode. If this were a real application, I'd set up
some sort of permissions system or have a separate administrator login.

For demonstration purposes though, I chose to implement the
administrator system this way.